Migrate ULP tests
